get pp e2e test working
