+++ ocf_datapipes/load/gsp.py
@@ -1,4 +1,5 @@
 import datetime
+import logging
 from pathlib import Path
 from typing import Optional, Union
 
@@ -8,6 +9,8 @@ import pandas as pd
 import xarray as xr
 from torchdata.datapipes import functional_datapipe
 from torchdata.datapipes.iter import IterDataPipe
+
+logger = logging.getLogger(__name__)
 
 try:
     from ocf_datapipes.utils.eso import get_gsp_metadata_from_eso, get_gsp_shape_from_eso
@@ -48,6 +51,8 @@ class OpenGSPIterDataPipe(IterDataPipe):
         )
         self._gsp_id_to_shape = gsp_id_to_shape  # Save, mostly for plotting to check all is fine!
 
+        logger.debug('Getting GSP data')
+
         # Load GSP generation xr.Dataset:
         gsp_pv_power_mw_ds = xr.open_dataset(self.gsp_pv_power_zarr_path, engine="zarr")
 
@@ -69,6 +74,7 @@ class OpenGSPIterDataPipe(IterDataPipe):
             y_osgb=gsp_id_to_shape.geometry.centroid.y.astype(np.float32),
             capacity_mwp=gsp_pv_power_mw_ds.installedcapacity_mwp.data.astype(np.float32),
         )
+
 
         del gsp_id_to_shape, gsp_pv_power_mw_ds
         while True:
