+++ ocf_datapipes/load/pv/live.py
@@ -70,8 +70,9 @@ class OpenPVFromDBIterDataPipe(IterDataPipe):
         logger.debug(f"There are now {len(pv_metadata.index)} pv system in the metadata")
         logger.debug(f"There are now {len(pv_power.columns)} pv system in the power data")
 
+
         data_xr = put_pv_data_into_an_xr_dataarray(
-            pv_power_watts=pv_power,
+            pv_power_watts=pv_power, # TODO check this is watts
             y_osgb=pv_metadata.y_osgb.astype(np.float32),
             x_osgb=pv_metadata.x_osgb.astype(np.float32),
             capacity_wp=pv_metadata.capacity_wp,
@@ -138,6 +139,9 @@ def get_metadata_from_database(providers: List[str] = None) -> pd.DataFrame:
 
     pv_system_all_df["capacity_kw"] = pv_system_all_df["installed_capacity_kw"]
     pv_system_all_df["capacity_wp"] = pv_system_all_df["capacity_kw"] * 1000
+
+    # sort index
+    pv_system_all_df = pv_system_all_df.sort_index()
 
     return pv_system_all_df
