+++ ocf_datapipes/select/location_picker.py
@@ -1,8 +1,12 @@
 import numpy as np
+import logging
 from torchdata.datapipes import functional_datapipe
 from torchdata.datapipes.iter import IterDataPipe
 
 from ocf_datapipes.utils.consts import Location
+
+
+logger = logging.getLogger(__name__)
 
 
 @functional_datapipe("location_picker")
@@ -24,6 +28,9 @@ class LocationPickerIterDataPipe(IterDataPipe):
     def __iter__(self) -> Location:
         """Returns locations from the inputs datapipe"""
         for xr_dataset in self.source_datapipe:
+
+            logger.debug(f'Getting locations for {xr_dataset}')
+
             if self.return_all_locations:
                 # Iterate through all locations in dataset
                 for location_idx in range(len(xr_dataset["x_osgb"])):
@@ -31,6 +38,7 @@ class LocationPickerIterDataPipe(IterDataPipe):
                         x=xr_dataset["x_osgb"][location_idx].values,
                         y=xr_dataset["y_osgb"][location_idx].values,
                     )
+                    logger.debug(f'Got all locations {location}')
                     yield location
             else:
                 # Assumes all datasets have osgb coordinates for selecting locations
